Add parsePseudoBigInt
