+++ Cargo.toml
@@ -11,6 +11,7 @@ repository = "https://github.com/yever/rusty-typescript"
 wasm-bindgen = "0.2"
 num-traits = "0.2"
 num-derive = "0.2"
+num-bigint = "0.2"
 js-sys = "0.3"
 snafu = "0.2"
