+++ TypeScript-replace/src/compiler/utilities.ts
@@ -8639,70 +8639,7 @@ namespace ts {
      * to its decimal string representation, e.g. `4660`.
      */
     export function parsePseudoBigInt(stringValue: string): string {
-        let log2Base: number;
-        switch (stringValue.charCodeAt(1)) { // "x" in "0x123"
-            case CharacterCodes.b:
-            case CharacterCodes.B: // 0b or 0B
-                log2Base = 1;
-                break;
-            case CharacterCodes.o:
-            case CharacterCodes.O: // 0o or 0O
-                log2Base = 3;
-                break;
-            case CharacterCodes.x:
-            case CharacterCodes.X: // 0x or 0X
-                log2Base = 4;
-                break;
-            default: // already in decimal; omit trailing "n"
-                const nIndex = stringValue.length - 1;
-                // Skip leading 0s
-                let nonZeroStart = 0;
-                while (stringValue.charCodeAt(nonZeroStart) === CharacterCodes._0) {
-                    nonZeroStart++;
-                }
-                return stringValue.slice(nonZeroStart, nIndex) || "0";
-        }
-
-        // Omit leading "0b", "0o", or "0x", and trailing "n"
-        const startIndex = 2, endIndex = stringValue.length - 1;
-        const bitsNeeded = (endIndex - startIndex) * log2Base;
-        // Stores the value specified by the string as a LE array of 16-bit integers
-        // using Uint16 instead of Uint32 so combining steps can use bitwise operators
-        const segments = new Uint16Array((bitsNeeded >>> 4) + (bitsNeeded & 15 ? 1 : 0));
-        // Add the digits, one at a time
-        for (let i = endIndex - 1, bitOffset = 0; i >= startIndex; i-- , bitOffset += log2Base) {
-            const segment = bitOffset >>> 4;
-            const digitChar = stringValue.charCodeAt(i);
-            // Find character range: 0-9 < A-F < a-f
-            const digit = digitChar <= CharacterCodes._9
-                ? digitChar - CharacterCodes._0
-                : 10 + digitChar -
-                (digitChar <= CharacterCodes.F ? CharacterCodes.A : CharacterCodes.a);
-            const shiftedDigit = digit << (bitOffset & 15);
-            segments[segment] |= shiftedDigit;
-            const residual = shiftedDigit >>> 16;
-            if (residual) segments[segment + 1] |= residual; // overflows segment
-        }
-        // Repeatedly divide segments by 10 and add remainder to base10Value
-        let base10Value = "";
-        let firstNonzeroSegment = segments.length - 1;
-        let segmentsRemaining = true;
-        while (segmentsRemaining) {
-            let mod10 = 0;
-            segmentsRemaining = false;
-            for (let segment = firstNonzeroSegment; segment >= 0; segment--) {
-                const newSegment = mod10 << 16 | segments[segment];
-                const segmentValue = (newSegment / 10) | 0;
-                segments[segment] = segmentValue;
-                mod10 = newSegment - segmentValue * 10;
-                if (segmentValue && !segmentsRemaining) {
-                    firstNonzeroSegment = segment;
-                    segmentsRemaining = true;
-                }
-            }
-            base10Value = mod10 + base10Value;
-        }
-        return base10Value;
+        return RustyTypeScript.parsePseudoBigInt(stringValue);
     }
 
     export function pseudoBigIntToString(pseudoBigInt: PseudoBigInt): string {
