+++ src/compiler/utilities.rs
@@ -1,14 +1,9 @@
 use crate::types::PseudoBigInt;
 use crate::types::PseudoBigIntJs;
+use num_bigint::BigUint;
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen(js_name = isPinnedComment)]
 pub fn is_pinned_comment(text: &str, start: usize) -> bool {
     text.chars().nth(start + 1) == Some('*') && text.chars().nth(start + 2) == Some('!')
-}
-
-#[wasm_bindgen(js_name = "pseudoBigIntToString")]
-pub fn pseudo_big_int_to_string(pseudo_big_int_js: &PseudoBigIntJs) -> String {
-    let pseudo_big_int: PseudoBigInt = pseudo_big_int_js.into();
-    pseudo_big_int.to_string()
 }
